requirements.txt
fastapi
uvicorn
jinja2
aiofiles
google-generativeai
python-dotenv
supabase
